Wide screen mode and zoom
Changes:
- implemented wide screen rendering (no more stretching)
- implemented zoom in the game, like in XBLA version
- fixed fades (running fast and no more flicker)
- stabilized the game FPS
- added FPS display indicator, and a way to toggle it in video options
- added sound effects to replays (e.g. ball kicks)
- fixed annoyingly bright pitches
- fixed invalid transparency in player shirts
- implemented stadium menu sprites properly
- fixed mouse not working in replay/exit menu after the match
- restored blocking initial mouse selection when game starts
- fixed menu background not reloading properly when asset resolution changes
- fixed menu item renderer constantly emptying the cache after the screen
  resolution changes
- implemented menu local sprites registration and cleanup, and fixed their
  rendering
- added missing playCrowdNoise() call to the start of replays
- fixed video options of Android showing up on PC when returning from window
  mode menu
- added option to use nearest neighbor scaling when converting SWOS assets
- removed game time, player name and result sprites, they're now rendered via
  direct calls (which simplified the code vastly)
- switched camera coordinates from fixed to floating point
- removal of more SWOS code and refactoring

+++ src/game/bench/bench.h
@@ -14,6 +14,6 @@ int getOpponentBenchY();
 void swapBenchWithOpponent();
 bool getBenchOff();
 void setBenchOff();
-int benchCameraX();
+float benchCameraX();
 const PlayerGame& getBenchPlayer(int index);
 int getBenchPlayerPosition(int index);
